requirements.txt
Python 3.6
selenium-3.141.0
pandas-0.25.1